refactor: normalize manager/service naming for people and voice
