+++ src/yoyopod/people/__init__.py
@@ -1,6 +1,6 @@
 """Mutable people-data seams for the YoyoPod runtime."""
 
-from yoyopod.people.directory import PeopleDirectory
+from yoyopod.people.manager import PeopleDirectory, PeopleManager
 from yoyopod.people.cloud_sync import build_cloud_contact
 from yoyopod.people.models import Contact, contacts_from_mapping, contacts_to_mapping
 
@@ -8,6 +8,7 @@ __all__ = [
     "build_cloud_contact",
     "Contact",
     "PeopleDirectory",
+    "PeopleManager",
     "contacts_from_mapping",
     "contacts_to_mapping",
 ]
